backend/rag-pipeline/requirements.txt
fastapi
uvicorn[standard]
apify-client
python-dotenv
requests
